Avoid glyph rendering with FT when not needed
If we only need to access the font metrics (like in scenegraph) for
layout, we don't need to render glyphs with FreeType at all.

Increase cached linearAdvance to 22 bits so that no overflow will
happen: FreeType returns 16.16 fixed point linearHoriAdvance, but
QFixed is 26.6, we store

    glyph->linearAdvance = linearHoriAdvance >> 10

Apparently 'short' is not enough since it's only 16 bits.

Change-Id: Id14eafa19f01a687de11997526281f9e7e860482
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@nokia.com>