Windows: Fix QSpinBox background setting via stylesheet
When using stylesheets, palette mask for drawing the background
of line edit child of spinbox is already correctly resolved to style
options in QStyleSheetStyle::drawPrimitive(), so we cannot simply
ignore that mask and check the palette mask of the parent spinbox.

Fixed by using a union mask of the parent spinbox palette and the
palette supplied by style option instead of simply using the mask from
parent spinbox. If either specifies custom base color, use that to
paint the background of the spinbox's line edit.

Task-number: QTBUG-24323
Change-Id: I1e738192db83b16d9bd48da54d29779e18788ef7
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@nokia.com>