Optimize HB memory consumption of PairPos1 in GPOS tables
This is pretty much the same fix as in
be0dfa34737a014e7ad060a90c4c24be9998fe13 but for PairPos
format 1 instead of 2.

With very simple GPOS tables we would waste a lot of memory on
caching an uncompressed table, so we now compress it in memory
as well.

Change-Id: I601331b4b83f636dab9e1ac403b343558c15b0de
Reviewed-by: Lars Knoll <lars.knoll@nokia.com>